pyproject.toml
[project]
name = "dataset-weighting-nn"
version = "0.1.0"
requires-python = ">=3.9,<3.12"
dependencies = [
  "numpy>=1.26",
  "pyyaml>=6.0",
  "tqdm>=4.66",
  "pillow>=10.0",
  "torch>=2.1",
  "boto3",
  "botocore",
  "matplotlib"
]

[project.optional-dependencies]
faiss = [
  "faiss-cpu>=1.7.4; platform_system == 'Linux' and python_version < '3.12'",
]
faiss-gpu = [
  "faiss-gpu==1.7.2; platform_system == 'Linux' and python_version < '3.12'",
]

[project.scripts]
dw-extract-embeddings = "dw.cli.extract_embeddings:main"
dw-retrieve-1nn = "dw.cli.retrieval_1nn:main"
dw-aggregate-weights = "dw.cli.aggregate_weights:main"


[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]